requirements.txt
playwright
python-dotenv
requests
jsonschema
openai

yapf
isort
pylint
